+++ CHANGELOG.md
@@ -6,12 +6,13 @@
 
 ### Добавлено
 
-- 
+- Coefficient calculating for RateInfo model [@fringer2423](https://github.com/fringer2423)
 
 ### Изменено
 
 - Refactor frontend [@fringer2423](https://github.com/fringer2423)
 - Frontend import paths
+- RateInfo model
 
 ## [0.0.32] - 2022-11-18